+++ prettier.config.cjs
@@ -1,0 +1,11 @@
+/* eslint-disable no-undef */
+module.exports = {
+	plugins: [require('prettier-plugin-tailwindcss')],
+	tailwindConfig: './tailwind.config.js',
+	useTabs: true,
+	printWidth: 80,
+	singleQuote: true,
+	trailingComma: 'none',
+	semi: false,
+	endOfLine: 'lf'
+}