+++ src/components/ChartComponents/ChartImages.vue
@@ -1,4 +1,3 @@
-<!-- eslint-disable @typescript-eslint/no-non-null-assertion -->
 <script setup lang="ts">
 import type { AlbumTile, DragDataTransfer, IndicesObject } from '#types'
 import { GlobalChartState } from '#shared/globals'
@@ -10,22 +9,26 @@ import {
 	getAlbumNumber
 } from '#shared/misc'
 
-import Dialog from '../CoreComponents/Dialog.vue'
-import Tooltip from '../CoreComponents/Tooltip.vue'
+import Dialog from '#core/Dialog.vue'
+import Tooltip from '#core/Tooltip.vue'
 
 const dialogId = 'DragLongHoldId'
 let selectedAlbumIndices: IndicesObject
 
 function openDialog(indices?: IndicesObject) {
-	// prettier-ignore
-	(document.getElementById(dialogId) as HTMLDialogElement).showModal()
+	const dialog = document.getElementById(dialogId) as HTMLDialogElement
+	dialog.showModal()
 
-	selectedAlbumIndices = indices!
+	if (!indices) {
+		return console.error('Error indices not defined in openDialog()')
+	}
+
+	selectedAlbumIndices = indices
 }
 
 function closeDialog() {
-	// prettier-ignore
-	(document.getElementById(dialogId) as HTMLDialogElement).close()
+	const dialog = document.getElementById(dialogId) as HTMLDialogElement
+	dialog.close()
 }
 
 function deleteSelectedAlbum() {
@@ -34,12 +37,23 @@ function deleteSelectedAlbum() {
 }
 
 function onDragOver(dragEvent: DragEvent) {
-	dragEvent.dataTransfer!.dropEffect = 'move'
+	if (!dragEvent.dataTransfer) {
+		return console.error(
+			'Error dragEvent.dataTransfer not defined in onDragOver',
+			dragEvent
+		)
+	}
+
+	dragEvent.dataTransfer.dropEffect = 'move'
 }
 
 function onDrop(dragEvent: DragEvent, { index1, index2 }: IndicesObject) {
-	// eslint-disable-next-line @typescript-eslint/no-non-null-asserted-optional-chain
-	const data = dragEvent.dataTransfer?.getData('text/plain')!
+	const data = dragEvent.dataTransfer?.getData('text/plain')
+
+	if (!data) {
+		return console.error('onDrop() failed: ', dragEvent, index1, index2, data)
+	}
+
 	const albumDraggedIn = JSON.parse(data) as DragDataTransfer
 
 	if (albumDraggedIn.dragSource === 'Chart') {
@@ -83,7 +97,12 @@ function onDragStart(dragEvent: DragEvent, { index1, index2 }: IndicesObject) {
 			index2: index2
 		}
 	})
-	dragEvent.dataTransfer!.dropEffect = 'copy'
+
+	if (!dragEvent.dataTransfer) {
+		return console.error('Error dragEvent.dataTransfer not defined in onDrop()')
+	}
+
+	dragEvent.dataTransfer.dropEffect = 'copy'
 }
 
 function deleteCurrent(indices: IndicesObject) {
@@ -138,7 +157,7 @@ function chartTitle(
 					:src="`${album.image}`"
 					:alt="'placeholder square'"
 					draggable="false"
-					class="select-none tw-chart-image-size"
+					class="tw-chart-image-size select-none"
 					@dragstart="() => undefined"
 					@dragover.prevent="() => undefined"
 					@drop.prevent="
@@ -156,9 +175,9 @@ function chartTitle(
 					<template #content>
 						<div class="tw-album-image-div-wrapper">
 							<img
-								v-show="album"
+								v-show="album && !GlobalChartState.options.lockChart"
 								src="/blackClose.svg"
-								class="hidden absolute m-1 left-0 top-0 group-hover:bg-white group-hover:block cursor-pointer"
+								class="absolute left-0 top-0 m-1 hidden cursor-pointer group-hover:block group-hover:bg-white"
 								@click="deleteCurrent({ index1, index2 })"
 							/>
 
@@ -167,8 +186,11 @@ function chartTitle(
 								:secondIndex="index2"
 								:src="`${album.image}`"
 								:alt="`${album.artist} - ${album.name}`"
-								class="select-none cursor-grab tw-chart-image-size"
-								draggable="true"
+								class="tw-chart-image-size select-none"
+								:class="{
+									'cursor-grab': !GlobalChartState.options.lockChart
+								}"
+								:draggable="GlobalChartState.options.lockChart ? false : true"
 								@dragstart="
 									(dragEvent) =>
 										onDragStart(dragEvent, { index1: index1, index2: index2 })
@@ -179,7 +201,11 @@ function chartTitle(
 										onDrop(dragEvent, { index1: index1, index2: index2 })
 								"
 								@touchstart.prevent="
-									(touchEvent) =>
+									(touchEvent) => {
+										if (GlobalChartState.options.lockChart) {
+											return
+										}
+
 										onTouchStart(
 											touchEvent,
 											album,
@@ -187,6 +213,7 @@ function chartTitle(
 											{ index1, index2 },
 											openDialog
 										)
+									}
 								"
 							/>
 							<div