+++ src/components/ChartComponents/ChartTitles.vue
@@ -1,4 +1,3 @@
-<!-- eslint-disable @typescript-eslint/no-non-null-assertion -->
 <script setup lang="ts">
 import { GlobalChartState } from '#shared/globals'
 import { getAlbumNumber, GrayBoxImgForPlaceholder } from '#shared/misc'
@@ -19,7 +18,7 @@ function albumNameEdited(event: Event, index: number, index2: number) {
 <template>
 	<div
 		v-if="GlobalChartState?.options.displayTitles"
-		class="text-left min-w-[200px] text-sm"
+		class="min-w-[200px] text-left text-sm"
 	>
 		<div
 			v-for="(albumRow, index) in GlobalChartState.chartTiles"